Do not show user API keys
This updates the detail and edit view of pipelines to not show the API
key of the user that connects with the Archivematica pipeline. The
current plain text field has been replaced with a password field to do
so.

The user edit view has been updated to have a Regenerate API key
checkbox similar to the one in the Dashboard which replaces the "if
you change your password, your API key will also be changed to a new,
randomly generated one" approach.

Finally the Storage Service user views and their redirects have been
updated to show the API key only once on re-generation. This uses
Django's messages framework to render an alert with a button to copy
the API key to the browser's clipboard.