Add workflow to build docker image and push to ECR
We are working on setting up automated deployment infrastructure for our
Archivematica instances. The deployment task (which will live in our
infrastructure repo) will need Docker images for each Archivematica
component to deploy. This commit adds a build action to create an
image for the storage service and upload it to ECR, where it will be
accessible to the deploy job.