Make targe repo of build action configurable
The build action in this repo is intended to be callable
from our infrastructure repo as part of an Archivematica
deploy. However, when a workflow is called from a separate
repo, it runs in the context of that repo, so by default
actions/checkout checks out the repo calling the workflow.
In this case we want it to actually check out this repo.
To achieve that, we make target_repo and input of the build
workflow, defaulting to this repo, and use target_repo to
explicitly specify what actions/checkout checks out.