Add API endpoint for reviewing AIP deletion requests
This change refactors the package request handling logic by extracting
it from the views and API resources modules within the locations
package into a dedicated module, enabling it to be shared across
views, API endpoints, and management commands.

Additionally, it introduces a new API endpoint in the package resource
that allows authorized users to approve or reject package deletion
requests. Access to this endpoint is restricted by the same permission
used in the “View delete requests” page of the user interface.